Path Combat accumulating for Red Correction
Path cards that have a predefined PathCombat defense on the card are incorrectly accumulating for Red.  

The tokens are now accumulating correcting but that updated had a bug that allowed PathCombat Defense to contribute for Red as well.  Since no PathCombat values for Red exist on Paths, this should be the final correction to resolve the Path Red Attack Token issue.